requirements.txt
matplotlib==2.0.0
numpy==1.18
scipy==1.1.0
tqdm==4.48.2
pillow
scikit-learn==0.21.3
opencv-python
numba==0.50.0
imageio